requirements.txt
awscli==1.11.13
blinker==1.3
botocore==1.4.70
chardet==2.3.0
cloud-init==18.5
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
cryptography==1.2.3
decorator==4.0.6
Django==2.2.2
djangorestframework==3.9.4
docutils==0.12
gunicorn==20.1.0
idna==2.0
ipython==2.4.1
Jinja2==2.8
jmespath==0.9.0
jsonpatch==1.10
jsonpointer==1.9
language-selector==0.1
linecache2==1.0.0
MarkupSafe==0.23
oauthlib==1.0.3
Pillow==3.1.2
prettytable==0.7.2
pyasn1==0.1.9
pycurl==7.43.0
Pygments==2.1
pygobject==3.20.0
PyJWT==1.3.0
pyserial==3.0.1
python-apt==1.1.0b1+ubuntu0.16.4.4
python-dateutil==2.4.2
python-debian==0.1.27
python-decouple==3.7
python-systemd==231
pytz==2019.1
PyYAML==3.11
requests==2.9.1
roman==2.0.0
rsa==3.2.3
s3transfer==0.1.9
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.3.0
ssh-import-id==5.5
testresources==1.0.0
traceback2==1.4.0
ufw==0.35
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.13.1
